refactor: organize hexagon section
