+++ ts/logic/hexagonSection.ts
@@ -1,3 +1,5 @@
+import { assertNonNullish } from "./utils";
+
 export function generateHexagon(hexagonNumber: number, className?: string) {
   return `<div class="hexagon__outer${className ? " " + className : ""}">
 	  <div class="hexagon__inner">${hexagonNumber}</div>
@@ -56,4 +58,33 @@ export function generateHexagons(
   }
 
   return html;
+}
+
+export function generateHexagonSection() {
+  const hexagonContainer = document.getElementById("hexagon__container");
+  assertNonNullish(
+    hexagonContainer,
+    "HTMLElement #hexagonContainer not found!"
+  );
+
+  const numberOfHexagonsElement = document.getElementById(
+    "hexagon-amount"
+  ) as HTMLInputElement;
+  assertNonNullish(
+    numberOfHexagonsElement,
+    "HTMLElement #numberOfHexagonsElement not found!"
+  );
+
+  const hexagonsFirstRowElement = document.getElementById(
+    "hexagon-first-row"
+  ) as HTMLInputElement;
+  assertNonNullish(
+    hexagonsFirstRowElement,
+    "HTMLElement #hexagonsFirstRowElement not found!"
+  );
+
+  hexagonContainer.innerHTML = generateHexagons(
+    Number(hexagonsFirstRowElement.value),
+    Number(numberOfHexagonsElement.value)
+  );
 }
